Array

Description

Arrays

Stuff to make it easier to work with arrays from the array package, especially when it comes to unboxed arrays. If the vector package is available, just use that.

Documentation

class Unbox a where

Minimal complete definition

Nothing

Associated Types

type Unboxed a

Methods

toU :: a -> Unboxed a

default toU :: Coercible a (Unboxed a) => a -> Unboxed a

frU :: Unboxed a -> a

default frU :: Coercible (Unboxed a) a => Unboxed a -> a

data Arr arr i a

Array type for element types isomorphic to other element types with existing array support. Primarily intended as a way to get unboxed arrays. As an example, if you have "newtype N = N Int", define "instance Unbox N where type Unboxed N = Int" and use UArr i N as an unboxed array for N. This is a lot more performant than a boxed Array i N. Also works with mutable arrays, IOUArr and STUArr.

type UArr = Arr UArray

type IOUArr = Arr IOUArray

type STUArr s = Arr (STUArray s)

data Arr2 arra arrb i ab

Arrays for 2-tuples Primarily intended as a way to get unboxed arrays. If a and b can be put in UArrays, UArr2 i (a,b) works as an unboxed array for (a,b). This is a lot more performant than a boxed Array i (a,b). Also works with mutable arrays, STUArr2 and IOUArr2. Can be nested to get UArr3, UArr4, etc. Use with Unbox and Arr to store your own types.

type UArr2 = Arr2 UArray UArray

type UArr3 = Arr2 UArray UArr2

type UArr4 = Arr2 UArr2 UArr2

type STUArr2 s = Arr2 (STUArray s) (STUArray s)

type STUArr3 s = Arr2 (STUArray s) (STUArr2 s)

type STUArr4 s = Arr2 (STUArr2 s) (STUArr2 s)

type IOUArr2 = Arr2 IOUArray IOUArray

type IOUArr3 = Arr2 IOUArray IOUArr2

type IOUArr4 = Arr2 IOUArr2 IOUArr2