Fenwick

Description

Fenwick tree, or binary indexed tree

A data structure supporting point updates and range queries, or the opposite. Large ranges, beyond typical memory limits, are supported.

Sources:

https://en.wikipedia.org/wiki/Fenwick_tree

Peter M. Fenwick, "A New Data Structure for Cumulative Frequency Tables", 1994 https://citeseerx.ist.psu.edu/viewdoc/summary?doi=10.1.1.14.8917

https://hackage.haskell.org/package/binary-indexed-tree

4 / / 2 6 1 3 5 7

Let n = r - l + 1 where (l, r) is the range of the Fenwick tree. The complexities assume (<>) takes O(1) time.

Documentation

data FTree a

emptyF :: Monoid a => (Int, Int) -> FTree a

Builds a Fenwick tree on range (l, r) where each element is mempty. O(log n).

fromListF :: Monoid a => (Int, Int) -> [a] -> FTree a

Builds a Fenwick tree on (l, r) where the elements are taken from a list. If the list is shorter than the range, the remaining elements are mempty. O(n).

mappendF :: Commutative a => Int -> a -> FTree a -> FTree a

mappends to the element at an index. O(log n).

foldPrefixF :: Monoid a => Int -> FTree a -> a

The result of folding the prefix upto the given index. Indices outside the tree range are allowed, it is assumed elements there are mempty. O(log n).

foldRangeF :: (Commutative a, Group a) => Int -> Int -> FTree a -> a

Folds the elements in the range (l, r). O(log n).

mappendRangeF :: (Commutative a, Group a) => Int -> Int -> a -> FTree a -> FTree a

mappends to all elements in the range (l, r). Can be used with foldPrefixF for point queries. O(log n).

binSearchF :: Monoid a => (a -> Bool) -> FTree a -> Maybe (Int, a)

Binary searches for the shortest prefix such that the fold of all values in it satisfies the given monotonic predicate. Returns the end index and the fold of the found prefix. O(log n).

toScanl1F :: Monoid a => FTree a -> [a]

Converts to a list of prefix accumulated values. O(n).