FloydWarshall

Description

Floyd-Warshall algorithm

Finds shortest paths between all pairs of vertices in a directed weighted graph.

Sources:

https://en.wikipedia.org/wiki/Floyd%E2%80%93Warshall_algorithm

Robert W. Floyd, "Algorithm 97: Shortest Path", Communications of the ACM, Vol. 5, No. 6, 1962 https://dl.acm.org/doi/10.1145/367766.368168

Documentation

type Weight = Int

data WEdge

Constructors

WEdge !Vertex !Vertex !Weight

floydWarshallFromEdges :: Bounds -> [WEdge] -> Vertex -> Vertex -> Maybe Weight

Runs the algorithm on the graph represented by the given list of edges. There should be no negative cycles. O(|V|^3). Queries take O(1).

floydWarshall :: MArray a Weight m => a (Vertex, Vertex) Weight -> m ()

Runs the algorithm on the adjacency matrix of the graph given as a mutable array. O(|V|^3).