HLD

Description

Heavy-light decomposition

Decomposition of a tree of size n into multiple paths, such that there are O(log n) paths between any pair of vertices.

HLD is useful for problems that have queries or updates on the path between two vertices.

This implementation maps each node to an integer position in [1..n] such that nodes in a path have consecutive positions. A path between two vertices decomposes into O(log n) ranges of consecutive integers, possibly simplifying the problem.

Sources:

https://en.wikipedia.org/wiki/Heavy_path_decomposition

https://cp-algorithms.com/graph/hld.html

Documentation

data HLD

Constructors

HLD

Fields

par_ :: !(UArray Vertex Vertex)

dep_ :: !(UArray Vertex Int)

hed_ :: !(UArray Vertex Vertex)

pos_ :: !(UArray Vertex Int)

siz_ :: !(UArray Vertex Int)

buildHLD :: Bounds -> Tree Vertex -> HLD

Builds the HLD structure from a tree. O(n).

posHLD :: HLD -> Vertex -> Int

The position for the given vertex. O(1).

pathHLD :: HLD -> Vertex -> Vertex -> [(Int, Int)]

A list of position ranges which make up the path from u to v. O(log n).

edgePathHLD :: HLD -> Vertex -> Vertex -> [(Int, Int)]

pathsHLD but excludes the LCA. Useful when working with edges, each edge can be mapped to the node it leads down into. O(log n).

subtreeHLD :: HLD -> Vertex -> (Int, Int)

A position range covering the subtree of the given node. O(1).

lcaHLD :: HLD -> Vertex -> Vertex -> Vertex

The lowest common ancestor of two vertices. O(log n).