KMP

Description

Knuth-Morris-Pratt algorithm

A string matching algorithm that generates a prefix function p from an input string s, where p(i) is the length of the longest prefix of s that ends at index i, excluding the prefix [0..i].

Sources:

https://en.wikipedia.org/wiki/Knuth%E2%80%93Morris%E2%80%93Pratt_algorithm

https://cp-algorithms.com/string/prefix-function.html

Documentation

prefixFunc :: Eq a => Int -> (Int -> a) -> UArray Int Int

Constructs the prefix function. The input sequence should be 0-indexed. O(n).

prefixFuncBS :: ByteString -> UArray Int Int

prefixFunc for a ByteString. O(n).