Kruskal

Description

Kruskal's algorithm

An algorithm to find the minimum spanning forest of an edge-weighted graph.

Sources

https://en.wikipedia.org/wiki/Kruskal%27s_algorithm

Joseph B. Kruskal, "On the shortest spanning subtree of a graph and the traveling salesman problem", 1956 https://www.ams.org/journals/proc/1956-007-01/S0002-9939-1956-0078686-7/

Documentation

type Weight = Int

data WEdge

Constructors

WEdge

Fields

getU :: !Vertex

getV :: !Vertex

getW :: !Weight

kruskal :: Bounds -> [WEdge] -> [WEdge]

Runs Kruskal's algorithm on the graph represented by the given list of edges. Returns the edges that are part of a minimum spanning forest. Vertices should be non-negative. O(|V| + |E|log|E|).