LCA

Description

Lowest common ancestor queries on a tree

Uses an Euler tour and a sparse table for range minimum queries, with an range size optimization from 2n to n, adapted from PyRival.

Sources:

https://en.wikipedia.org/wiki/Lowest_common_ancestor

Michael Bender and Martin Farach-Colton, "The LCA Problem Revisited", 2000 https://www.ics.uci.edu/~eppstein/261/BenFar-LCA-00.pdf

https://github.com/cheran-senthil/PyRival/blob/master/pyrival/graphs/lca.py

Implementation note: l - 1 is taken as a dummy root for a forest, converting it to a tree to make things simpler.

Documentation

data LCA

buildLCA :: Bounds -> Tree Vertex -> LCA

Build a structure for LCA queries on a tree. O(n log n).

queryLCA :: Vertex -> Vertex -> LCA -> Vertex

Query the LCA of two nodes in a tree. O(1).

build1LCA :: Bounds -> [Tree Vertex] -> LCA

Build a structure for LCA queries on a forest. O(n log n).

query1LCA :: Vertex -> Vertex -> LCA -> Maybe Vertex

Query the LCA of two nodes in a forest. O(1).