LabelledGraph

Description

Utilities for working with edge-labeled/weighted graphs

There are useful definitions and algorithms in Data.Tree and Data.Graph but sadly these only deal with unlabeled graphs. Most definitions here mirror those in Data.Graph.

Documentation

type LEdge b = (Vertex, (b, Vertex))

type LGraph b = Array Vertex [(b, Vertex)]

data LTree b a

Constructors

LNode

Fields

rootLabelL :: a

subForestL :: [(b, LTree b a)]

buildLG :: Bounds -> [LEdge b] -> LGraph b

Builds a LGraph from a list of LEdges. O(n + m) for bounds size n and m edges.

dfsLTree :: LGraph b -> Vertex -> LTree b Vertex

For a LGraph that is known to be a tree, returns its LTree representation. O(n).

lTreeToTree :: LTree b a -> Tree a

Drops labels from an LTree. O(n).