MInt

Description

Modular arithmetic

Instances of Eq, Num, Fractional exist for MInt. All the usual operations take O(1) time, except for recip which takes O(log m) time. An instance of Enum exists for MInt. The enum is cyclic, it wraps to 0 after m-1. An instance of IArray UArray MInt exists. Instances of MArray (STUArray s) MInt (ST s) and MArray IOUArray MInt IO are also defined.

Documentation

newtype MInt

Integer type for modular arithmetic, using fixed prime modulo.

Constructors

MInt

Fields

toInt :: Int

mkMInt :: Int -> MInt

Int to MInt. The MInt constructor can be used directly if the value is known to be in [0.. m-1]. O(1).

inv :: MInt -> MInt

Reciprocal of an MInt. O(log m).