Misc

Description

Miscellaneous functions/recipes

Documentation

pairs :: [a] -> [(a, a)]

Generates distinct pairs of elements from a list. O(n^2).

fArray :: (IArray a e, Ix i) => (i, i) -> (i -> e) -> a i e

Generates an Array from bounds and a function. O(n) assuming f takes O(1).

chunksOf :: Int -> [a] -> [[a]]

Splits a list into chunks of fixed size. O(n).

replicateL :: Int -> [a] -> [a]

Replicates a list n times. O(nm) where m is the length of the list.

unique :: Eq a => [a] -> [a]

Eliminates consecutive duplicate elements. O(n).

foldExclusive :: (b -> a -> b) -> b -> [a] -> [b]

Folds strictly such that the ith element of the output list contains the fold of all elements in the input list except for the ith element. The fold function f must be commutative, in the sense that (b f a1) f a2 = (b f a2) f a1 f is called O(n log n) times.

modifyArray :: (MArray a e m, Ix i) => a i e -> i -> (e -> e) -> m ()

Modifies an element in a mutable array.

modifyArray' :: (MArray a e m, Ix i) => a i e -> i -> (e -> e) -> m ()

Modifies an element in a mutable array. Strict version.

ifoldr :: Foldable f => (Int -> a -> b -> b) -> b -> f a -> b

foldr with an index, starting at 0.

ifoldl' :: Foldable f => (b -> Int -> a -> b) -> b -> f a -> b

foldl' with an index, starting at 0.

foldMComp :: Monad m => (b -> a -> b) -> (c -> m a) -> b -> c -> m b

Compose a strict left fold function with a mapM function. Useful for foldMs. foldM (f foldMComp g) z = fmap (foldl' f z) . mapM g

farthest :: (a -> Maybe a) -> a -> a

Repeatedly applies a function to a value and returns the value which gives back Nothing.

foldTree' :: (a -> b -> c) -> (c -> b -> b) -> b -> Tree a -> c

Folds a tree. This does the same job as Data.Tree.foldTree but with different fold functions, which may be preferable in cases such as CPS folds.

class Semigroup a => Commutative a

A semigroup where the operation (<>) is commutative. a <> b = b <> a

class Monoid a => Group a where

A monoid where elements have inverses. a <> invert a = mempty invert a <> a = mempty

Methods

invert :: a -> a

class Semigroup a => Idempotent a

A semigroup where the elements are idempotent. a <> a = a

class (Monoid u, Monoid a) => Action u a where

A right monoid action of u on a. (x act u1) act u2 = x act (u1 <> u2) x act mempty = x

Methods

act :: a -> u -> a

bitLength :: FiniteBits b => b -> Int

Returns the number of bits required to represent the value.

odds :: [a] -> [a]

The elements at odd positions of a list.

evens :: [a] -> [a]

The elements at even positions of a list.

orM :: Monad m => m Bool -> m Bool -> m Bool

Short-circuiting monadic ||.

andM :: Monad m => m Bool -> m Bool -> m Bool

Short-circuiting monadic &&.

anyM :: (Monad m, Foldable f) => (a -> m Bool) -> f a -> m Bool

Monadic version of any.

allM :: (Monad m, Foldable f) => (a -> m Bool) -> f a -> m Bool

Monadic version of all.