Mod

Description

Modular arithmetic

Mod m i is an integer of type i modulo m. m will usually be a compile-time constant. If m is not known at compile time, Mod can still be used via GHC.TypeNats.SomeNat. m must be >= 2 and fit in the type i. Integer operations on i for values in [0..m-1] must not overflow. Examples to avoid are using (-) with a word type or using (*) with a large enough mod that (m-1)^2 oveflows.

This is a very general type, for something simpler see MInt.hs.

Instances of Eq, Num, Fractional exist for Mod m i. All the usual operations take O(1) time, except for recip which takes O(log n) time. This assumes Integral i methods take O(1) time. An instance of Enum exists for MInt. The enum is cyclic, it wraps to 0 after m-1. An instance of IArray UArray MInt exists. Instances of MArray (STUArray s) MInt (ST s) and MArray IOUArray MInt IO are also defined.

M7 Commonly used modulus.

M3 Commonly used modulus.

Documentation

newtype Mod (m :: Nat) i

Type for modular arithmetic modulo m with underlying type i.

Constructors

Mod

Fields

unMod :: i

type M7 = Mod 1000000007 Int

type M3 = Mod 998244353 Int

invMaybe :: forall m i. (KnownNat m, Integral i) => Mod m i -> Maybe (Mod m i)

The multiplicative inverse modulo m. It exists if and only if the number is coprime to m. O(log n).