Prufer

Description

Prufer sequences

Functions to convert trees to Prufer sequences and vice versa

Sources:

https://en.wikipedia.org/wiki/Pr%C3%BCfer_sequence

Xiaodong Wang, Lei Wang, Yingjie Wu, "An Optimal Algorithm for Prufer Codes", 2009 https://www.scirp.org/pdf/JSEA20090200006_93737200.pdf

Implementation note: The c array tracks the number of children of a node.

Documentation

graphToSeq :: Graph -> [Vertex]

Convert a bidirected Graph to a Prufer sequence. The graph must be a connected tree or empty. O(n).

treeToSeq :: Bounds -> Tree Vertex -> [Vertex]

Convert a Tree to a Prufer sequence. The tree must be rooted at r. O(n).

seqToEdges :: Bounds -> [Vertex] -> [Edge]

Convert a Prufer sequence to a list of edges, each edge present twice in its two orientations. The sequence must be valid. O(n).

seqToGraph :: Bounds -> [Vertex] -> Graph

Convert a Prufer sequence to a bidirected Graph. The sequence must be valid. O(n).