SegTree

Description

Segment tree

A data structure supporting point updates and range queries on a sequence of monoids. This implementation supports large ranges that do not fit in memory, sometimes called a dynamic or sparse segment tree.

Sources:

https://cp-algorithms.com/data_structures/segment_tree.html

The complexities below assume mappend takes O(1) time. Let n = r - l + 1 in all instances below.

SegTree implementable Foldable. Folding over all the elements takes O(n).

Documentation

data SegTree a

emptyST :: Monoid a => (Int, Int) -> SegTree a

Builds a segment tree on range (l, r) where each element is mempty. O(log n).

fromListST :: Monoid a => (Int, Int) -> [a] -> SegTree a

Builds a segment tree on (l, r) where the elements are taken from a list. If the list is shorter than the range, the remaining elements are mempty. O(n).

adjustST :: Monoid a => (a -> a) -> Int -> SegTree a -> SegTree a

Adjusts the element at index i. O(log n).

foldRangeST :: Monoid a => Int -> Int -> SegTree a -> a

Folds the elements in the range (ql, qr). Elements outside (l, r) are considered to be mempty. O(log n).