SegTreeLazy

Description

Segment tree with lazy propagation

A data structure supporting point updates, range queries, and certain range updates on a sequence of monoids. This differs from an ordinary segment tree in its ability to apply updates on a range, they are otherwise identical. In fact, Segtree a can be defined as LazySegTree () a. This implementation, like SegTree, supports large ranges that may not fit in memory.

Sources:

https://cp-algorithms.com/data_structures/segment_tree.html

LazySegTree u a is a segment tree on elements of type a and updates of type u. a and u must be monoids. An instance of Action u a must exist, which specifies a (right) monoid action of u on a. The following laws hold for a monoid action:

(x act u1) act u2 = x act (u1 <> u2)

x act mempty = x The segment tree requires an additional law:

(x1 <> x2) act u = (x1 act u) <> (x2 act u)

The complexities below assume <> for u, <> for a and act all take O(1) time. Let n = r - l + 1 in all instances below.

emptyST Builds a segment tree on range (l, r) where each element is mempty. O(log n).

fromListST Builds a segment tree on (l, r) where the elements are taken from a list. If the list is shorter than the range, the remaining elements are mempty. O(n).

boundsST The bounds of the segment tree. O(1).

adjustST Adjusts the element at index i. O(log n).

foldRangeST Folds the elements in the range (ql, qr). Elements outside (l, r) are considered to be mempty. O(log n).

Documentation

data LazySegTree u a

class (Monoid u, Monoid a) => Action u a where

Methods

act :: a -> u -> a

emptyLST :: Action u a => (Int, Int) -> LazySegTree u a

fromListLST :: Action u a => (Int, Int) -> [a] -> LazySegTree u a

boundsLST :: LazySegTree u a -> (Int, Int)

adjustLST :: Action u a => (a -> a) -> Int -> LazySegTree u a -> LazySegTree u a

updateRangeLST :: Action u a => u -> Int -> Int -> LazySegTree u a -> LazySegTree u a

Applies an update on elements in the range (ql, qr). O(log n).

foldRangeLST :: Action u a => Int -> Int -> LazySegTree u a -> a

foldrLST :: Action u a => (a -> b -> b) -> b -> LazySegTree u a -> b

Right fold over the elements of the segment tree. O(n). LazySegTree u cannot be Foldable because of the Action constraint :(