ZFunc

Description

Z-function

A string matching algorithm. The Z-algorithm generates the function z from a string s, where z(i) is the length of the longest prefix of s that starts at i. z[0] is set to 0.

Sources:

https://cp-algorithms.com/string/z-function.html

Documentation

zFunc :: Eq a => Int -> (Int -> a) -> UArray Int Int

Constructs the Z-function. The input sequence should be 0-indexed. O(n).

zFuncBS :: ByteString -> UArray Int Int

zFunc for a ByteString. O(n).