Array

Description

Arrays

Array type for element types isomorphic to other element types with existing array support.

Primarily useful for unboxed arrays. As an example, if you have "newtype N = N Int", define "instance Unbox N where type Unboxed N = Int" and use UArr i N as an unboxed array for N. Also works with mutable arrays, IOUArr and STUArr.

Documentation

class Unbox a where

Minimal complete definition

Nothing

Associated Types

type Unboxed a

Methods

toU :: a -> Unboxed a

default toU :: Coercible a (Unboxed a) => a -> Unboxed a

frU :: Unboxed a -> a

default frU :: Coercible (Unboxed a) a => Unboxed a -> a

data Arr arr i a

type UArr = Arr UArray

type IOUArr = Arr IOUArray

type STUArr s = Arr (STUArray s)