Misc

Description

Miscellaneous functions/recipes

Documentation

pairs :: [a] -> [(a, a)]

Generates distinct pairs of elements from a list. O(n^2).

fArray :: (IArray a e, Ix i) => (i, i) -> (i -> e) -> a i e

Generates an Array from bounds and a function. O(n) assuming f takes O(1).

chunksOf :: Int -> [a] -> [[a]]

Splits a list into chunks of fixed size. O(n).

replicateL :: Int -> [a] -> [a]

Replicates a list n times. O(nm) where m is the length of the list.

unique :: Eq a => [a] -> [a]

Eliminates consecutive duplicate elements. O(n).

foldExclusive :: (b -> a -> b) -> b -> [a] -> [b]

Folds a list of values such that the ith element of the result contains the folded result of all elements in the input list excluding the ith element. The fold is strict. The elements get folded in a not-very-simple order, so the following should hold: (b f a1) f a2 = (b f a2) f a1 O(n log n) assuming f takes O(1).

modifyArray :: (MArray a e m, Ix i) => a i e -> i -> (e -> e) -> m ()

Modifies an element in a mutable array.

modifyArray' :: (MArray a e m, Ix i) => a i e -> i -> (e -> e) -> m ()

Modifies an element in a mutable array. Strict version.

foldMComp :: Monad m => (b -> a -> b) -> (c -> m a) -> b -> c -> m b

Compose a strict left fold function with a mapM function. Useful for foldMs. foldM (f foldMComp g) z = fmap (foldl' f z) . mapM g

farthest :: (a -> Maybe a) -> a -> a

Repeatedly applies a function to a value and returns the value which gives back Nothing.

foldTree' :: (a -> b -> c) -> (c -> b -> b) -> b -> Tree a -> c

Folds a tree. This does the same job as Data.Tree.foldTree but with different fold functions, which may be preferable in cases such as CPS folds.

class Semigroup a => Commutative a

A semigroup where the operation (<>) is commutative. a <> b = b <> a

class Monoid a => Group a where

A monoid where elements have inverses. a <> invert a = mempty invert a <> a = mempty

Methods

invert :: a -> a